GitHub issues reporting (#22)
* One last big push, then smaller fixes

Had this going already before the small push idea.
Fixed the p-8 issue
Added highlighter.js
Adjusted the icons on main page. (Right to left, more variability.)

* Report missing tokens to assets

Now will report missing icons to assets repo instead of Trust Wallet.
There is now a MD template that will pre-fill in the details of the missing token in an Issue.  User will need to be logged into Github, but does not need to be part of the repo.

---------

Co-authored-by: gibsfinancedev <162373480+gibsfinancedev@users.noreply.github.com>